% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countrymap.R
\name{countrymap}
\alias{countrymap}
\title{Country map}
\usage{
countrymap(currentyear, oldyear, title = "", save = FALSE,
  interactive = FALSE)
}
\arguments{
\item{currentyear}{The vector of a dataset for the most current year}

\item{oldyear}{The vector of a dataset for the year you want to compare to}

\item{title}{The title of the plot and the name that the plot is saved to}

\item{save}{TRUE to save, FALSE to not}

\item{interactive}{TRUE to choose the interactive mode in viewer, FALSE to view in just the default plot}
}
\value{
a s4 object that has the difference of the datasets mapped to it
}
\description{
Takes in two years worth of datasets and returns a map of the
international countries that maps the difference of the two datasets
}
\examples{
##Creating the map of the data with data from 2001 and 2000, as well as using save and interactive move
countrymap2000.2001 <- countrymap(yearsdata$X2001, yearsdata$X2000, title = "Change in Students 2000-2001",
save = TRUE, interactive = TRUE)


}
