% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaningData.R
\name{datatodf}
\alias{datatodf}
\title{Cleaning Data}
\usage{
datatodf(x, name)
}
\arguments{
\item{x}{Clean text files that have been parsed into a readable format by
the read years function}

\item{year}{The fall year of the dataset}
}
\value{
a dataframe for the dataset given
}
\description{
Reads in all the data in a clean format
}
