% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapstu-package.R
\docType{package}
\name{mapstu-package.R}
\alias{mapstu-package.R}
\title{The mapstu package}
\description{
\tabular{ll}{ Package \tab mapstu \cr Type: \tab Package\cr
Version: \tab 0.1-1\cr Date: \tab 2017-01-23\cr License: \tab GPL-3\cr }
}
\details{
The mapstu package uses the geographical distribution data of the entire
Williams College student population published in the Williams Course
Catalogs to compare which states, and countries students are from each
year. This is done for fifteen years from 2000 to 2015 by pulling data
from text versions of the geographical distribution data. The core
components of the package creates different cloropleth maps based on
the arguments passed to the function. With the visual support of
the cloropleth map, one can easily visualize the difference of
student geographical distribution between any two years of the data
set. A printed comparison of the two years can also produced.
}
