% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perchange.R
\name{perchange}
\alias{perchange}
\title{Percent Change}
\usage{
perchange(year1, year2)
}
\arguments{
\item{year1}{The data vector that you want to subtract from}

\item{year2}{The data vector is that will be subtracted from}
}
\value{
a dataframe of percent change for each country
}
\description{
Provides a dataframe of percent change that can also be mapped
    The percent change function ignores NaN values, and Inf Values.
}
