% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readYears.R
\name{readYears}
\alias{readYears}
\title{Reading the txt file}
\usage{
readYears(dataset, name)
}
\arguments{
\item{dataset}{A Geographical Distribution dataset}

\item{name}{The name the cleaned dataset is saved to}
}
\value{
A cleaned dataset
}
\description{
Takes in the William's Geographical Distribution data and returns a
clean format (Area, Number) that can be read into a data frame.
}
