% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totaldata.R
\name{totaldata}
\alias{totaldata}
\title{Cleaning Data}
\usage{
totaldata()
}
\arguments{
\item{x}{Clean text files that have been parsed into a readable format by
the read years function}

\item{year}{The fall year of the dataset}
}
\value{
a dataframe for the dataset given
}
\description{
Reads in all the data in a clean format
}
