% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usmap.R
\name{usmap}
\alias{usmap}
\title{StateRegions}
\usage{
usmap(currentyear, oldyear, namesave, title)
}
\arguments{
\item{currentyear}{The vector of a dataset for the most current year}

\item{oldyear}{The vector of a dataset for the year you want to compare to}

\item{title}{The title of the plot}

\item{name}{The name that the map should be saved to}
}
\value{
a s4 object that has the difference of the datasets mapped to it
}
\description{
Takes in two years worth of datasets and returns a map of the
US States that maps the difference of the two datasets
}
\examples{
##Creating the map of the US with data from 2001 and 2002
usmap2000.2002 <- usmap(yearsdata$X2001, yearsdata$X2000, "usmap2000.2001.png", "Change in Students 2000-2001")

}
