% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldplot.R
\name{worldplot}
\alias{worldplot}
\title{Worldplot}
\usage{
worldplot(year, title = "", US = TRUE, save = FALSE,
  interactive = FALSE)
}
\arguments{
\item{year}{The vector for the year that is to be mapped}

\item{title}{The title of the plot and the name which it will be saved to}

\item{US}{choice of mapping US territories or mapping international countries}

\item{save}{TRUE to save, FALSE to not}

\item{interactive}{TRUE to choose the interactive mode in viewer, FALSE to view in just the default plot}
}
\value{
a S4 object that has the year dataset mapped to it
}
\description{
Takes in a year vector of data and returns a map of the US or
international countries that maps the geographical distribution of the datasets
}
\examples{
##Creating a state map of the data with data from 2001
usmap2001 <- worldplot(yearsdata$X2001, title = "Change in Students 2000-2001")

##Creating an international map with interactive view
worldmap <- worldplot(yearsdata$X2014, title = "Change in Students 2000-2001", interactive = TRUE)

}
