% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{comparison}
\alias{comparison}
\title{Comparison}
\usage{
comparison(year1, year2)
}
\arguments{
\item{year1}{The data vector that you want to subtract from. It is the year that will be used first}

\item{year2}{The data vector is that will be subtracted from}
}
\value{
a comparison summary between year 1 and year 2
}
\description{
Provides comparison summary data of the students that were in each class.
    For example, the mean change of the number of students across all areas
    or the total number of students accepted each year, or the percent change of students
    from year to year. The code is still somewhat buggy as there are territories mentioned
    in the Williams data that are not official countries or states, such as Palestine territory occuppie.
    Thus the (#ofinternationals) + (#ofdomestics) do not always equal (#oftotalstudents).
}
