% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countrymap.R
\name{countrymap}
\alias{countrymap}
\title{CountryRegions}
\usage{
countrymap(currentyear, oldyear, namesave, title)
}
\arguments{
\item{currentyear}{The vector for the most current year}

\item{oldyear}{The vector for the year you want to compare to}

\item{title}{The title of the plot}

\item{name}{The name that the map should be saved to}
}
\value{
a s4 object that has the difference of the datasets mapped to it
}
\description{
Takes in two years worth of datasets and returns a map of the
international countries that maps the difference of the two datasets
}
\examples{
##Creating the map of the data with data from 2001 and 2000
countrymap2000.2001 <- countrymap(yearsdata$X2001, yearsdata$X2000, "countrymap2000.2001.png", "Change in Students 2000-2001")

}
