% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebyside.R
\name{sidebyside}
\alias{sidebyside}
\title{Sidebyside}
\usage{
sidebyside(year1, year2, title1 = "", title2 = "", savename = "plot",
  US = TRUE, save = FALSE, interactive = FALSE, change = FALSE)
}
\arguments{
\item{year1}{The vector for a year of data. It is important to note that
you can also plot sidebyside data of differences}

\item{year2}{The vector for a year of data}

\item{title1}{The title of the plot and the name which it will be saved to}

\item{title2}{The title of the second plot}

\item{savename}{The name the plot is saved to}

\item{US}{choice of mapping US territories or mapping international countries}

\item{save}{Whether you want to save or not}

\item{interactive}{Chooses the interactive mode in viewer, or just the default plot}
}
\value{
sidebyside S4 object that has the different year datasets mapped to it
}
\description{
Plots two of the years side by side to better compare the diversity of
Williams College's geographical distribution of students. There is an
international options, and a statemap option.
}
\examples{
##Creating a state map of the data with data from 2001
usmap2001 <- worldplot(yearsdata$X2001, title = "Change in Students 2000-2001")

##Creating an international map with interactive view
worldmap <- worldplot(yearsdata$X2014, title = "Change in Students 2000-2001", interactive = TRUE)

}
