% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usmap.R
\name{usmap}
\alias{usmap}
\title{StateRegions}
\usage{
usmap(currentyear, oldyear, namesave, title)
}
\arguments{
\item{currentyear}{The vector of a dataset for the most current year}

\item{oldyear}{The vector of a dataset for the year you want to compare to}

\item{title}{The title of the plot}

\item{name}{The name that the map should be saved to}
}
\value{
a s4 object that has the difference of the datasets mapped to it
}
\description{
Takes in two years worth of datasets and returns a map of the
US States that maps the difference of the two datasets
}
