% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usmap.R
\name{usmap}
\alias{usmap}
\title{US Map}
\usage{
usmap(currentyear, oldyear, title = "", save = FALSE, interactive = FALSE)
}
\arguments{
\item{currentyear}{The vector of a dataset for the most current year}

\item{oldyear}{The vector of a dataset for the year you want to compare to}

\item{title}{The title of the plot and the name which the plot is saved to}

\item{save}{TRUE to save, FALSE to not}

\item{interactive}{TRUE to choose the interactive mode in viewer, FALSE to view in just the default plot}
}
\value{
a s4 object that has the difference of the datasets mapped to it
}
\description{
Takes in two years worth of datasets and returns a map of the
US States that maps the difference of the two datasets
}
\examples{
##Creating the map of the US with data from 2001 and 2002
usmap2000.2002 <- usmap(yearsdata$X2001, yearsdata$X2000, title = "Change in Students 2000-2001")

}
